refactor: total-repair 관련 컴포넌트 파일 이동
