+++ src/templates/DashboardTemplate.jsx
@@ -3,10 +3,10 @@ import styled from 'styled-components';
 import { color, typo } from '../styles/tokens';
 import RepairStatusBox from '../components/dashboard/RepairStatusBox';
 import Calendar from '../components/dashboard/Calendar';
-import RepairDetailBox from '../components/common/BoxRepairDetail';
+import RepairDetailBox from '../components/total-repair/BoxRepairDetail';
 import { Column, Row } from '../styles/flex';
 import Modal from '../components/common/Modal';
-import ModalRepairDetail from '../components/common/ModalRepairDetail';
+import ModalRepairDetail from '../components/total-repair/ModalRepairDetail';
 import ModalImageSlider from '../components/common/ModalImageSlider';
 import { useNavigate } from 'react-router-dom';
 import { isToday, format } from 'date-fns';
